User token FK type

+++ database/migrations/2021_02_12_043152_create_table_user_token.php
@@ -14,7 +14,7 @@ class CreateTableUserToken extends Migration {
       Schema::create('user_token', function (Blueprint $table) {
          $table->uuid('id')->primary();
          // $table->uuid('user_id')->nullable(false);
-         $table->unsignedInteger('user_id')->nullable(false);
+         $table->unsignedBigInteger('user_id')->nullable(false);
          $table->string('token', 255)->unique();
          $table->dateTime('expire_date')->nullable(false);
          $table->string('user_agent')->nullable();
